Change met_barpres to air_pressure, because it is not corrected for sensor height
1. Change Display Name from "Barometric Pressure" to "Air Pressure"
2. Change Standard Name from "air_pressure_at_sea_level" to
"air_pressure"
3. Change Description from "Barometric Pressure is a measure of the
weight of the column of air above the sensor. It is also commonly
referred to as atmospheric pressure. This data products is calculated
by converting the measured Level 0 Barometric Pressure from mbar to Pa"
to "Air Pressure is a measure of the weight of the column of air above
the sensor, not adjusted for sensor height above sea level. This data
product is calculated by converting the measured Level 0 Barometric
Pressure from mbar to Pa."
4. Change Long Name from "Barometric Pressure" to "Air Pressure"