added a fix for CSPPs based on ticket 11056
The parameter pressure from the CSPPs were corrected as follow:

"name": "ctd_pressure"
"display_name": "CTD Seawater Pressure"
"description": "Seawater Pressure refers to the pressure exerted on a
sensor in situ by the weight of the column of seawater above it. It is
calculated by subtracting one standard atmosphere from the absolute
pressure at the sensor to remove the weight of the atmosphere on top of
the water column. The pressure reported is the CTD pressure recorded at
the time of the sensor reading."